# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APP API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/business-posts": {
            "put": {
                "summary": "Update business post by ID",
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessPostsController::update",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "id of business post",
                                        "type": "string"
                                    },
                                    "business_id": {
                                        "description": "id of business",
                                        "type": "string"
                                    },
                                    "text": {
                                        "description": "id of user",
                                        "type": "string"
                                    },
                                    "photo": {
                                        "description": "base64 encoded image",
                                        "type": "string"
                                    },
                                    "expire_date": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Business post updated"
                    },
                    "400": {
                        "description": "Invalid given data"
                    },
                    "404": {
                        "description": "Business post not found"
                    }
                }
            },
            "post": {
                "summary": "Create business post",
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessPostsController::store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "business_id": {
                                        "description": "id of business",
                                        "type": "string"
                                    },
                                    "text": {
                                        "description": "id of user",
                                        "type": "string"
                                    },
                                    "photo": {
                                        "description": "base64 encoded image",
                                        "type": "string"
                                    },
                                    "expire_date": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "BusinessPostResource"
                    }
                }
            },
            "delete": {
                "summary": "Delete a business post",
                "operationId": "App\\Http\\Controllers\\API\\v1\\BusinessPostsController::destroy",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "id of business post",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Business Post Resource"
                    }
                }
            }
        },
...
        "/api/v1/users": {
            "put": {
                "summary": "Update current logged in user",
                "operationId": "App\\Http\\Controllers\\API\\v1\\UsersController::update",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Update current logged in user",
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "description": "Update current logged in user",
                                        "type": "string"
                                    },
                                    "age_group": {
                                        "description": "Update current logged in user",
                                        "type": "string"
                                    },
                                    "gender": {
                                        "description": "Update current logged in user",
                                        "type": "string"
                                    },
                                    "cover_photo": {
                                        "description": "Update current logged in user",
                                        "type": "string"
                                    },
                                    "avatar_photo": {
                                        "description": "Update current logged in user",
                                        "type": "string"
                                    },
                                    "bio": {
                                        "description": "Update current logged in user",
                                        "type": "string"
                                    },
                                    "allow_location_tracking": {
                                        "description": "Update current logged in user",
                                        "type": "boolean"
                                    },
                                    "post_publicly": {
                                        "description": "Update current logged in user",
                                        "type": "boolean"
                                    },
                                    "t_c_agreed": {
                                        "description": "Update current logged in user",
                                        "type": "boolean"
                                    },
                                    "profile_visible": {
                                        "description": "Update current logged in user",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "UserResource"
                    }
                }
            },
            "delete": {
                "summary": "Delete current logged in user",
                "operationId": "App\\Http\\Controllers\\API\\v1\\UsersController::destroy",
                "responses": {
                    "200": {
                        "description": "bool"
                    }
                }
            }
        }
    }
}